+++ docs/guides/changelog.md
@@ -11,6 +11,7 @@ description: |-
 
 **Features:**
 * Add `azuresql_view` as a data source
+* Add `azuresql_database` as a resource
 * Add optional parameters `subscription_id`, `check_server_exists` and `check_database_exists` which can be used to check database/server existence before connecting.
 
 **Bugfixes:**